requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
jinja2==3.1.4
httpx==0.27.0
pillow==10.3.0
google-generativeai==0.7.2
python-dotenv==1.0.1
